Edit state links interface

+++ pages/api/article/delete.ts
@@ -2,6 +2,7 @@ import type { NextApiRequest, NextApiResponse } from "next";
 import connect from "lib/mongo";
 import Article from "models/Article";
 import admin from "lib/firebaseAdmin";
+import { getStateByCode } from "lib/localData";
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse<any>) {
   const token = req.headers.authorization;
@@ -18,7 +19,14 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse<
   }
 
   try {
+    const state = getStateByCode(article?.stateCode);
+    if (!state) {
+      throw new Error("Invalid state code");
+    }
+
     await Article.findByIdAndDelete(id);
+
+    await res.revalidate(`/${state.country}/${state.slug}`);
     res.status(200).json({ success: true });
   } catch (error: any) {
     res.status(500).json({ error: error.message });
